requirements.txt
streamlit==1.39.0
sentence-transformers==3.0.1
transformers==4.44.2
faiss-cpu==1.12.0
pypdf==6.1.0
torch==2.8.0 